requirements.txt
django>=4.0
mysqlclient
pillow
torch
transformers
requests
